Complete Phase 1: kmer pipeline optimizations
Extract kmers directly from ASCII sequence in a single pass
(kmers_from_ascii), eliminating the Read struct encoding/decoding
round-trip. Switch kmer counts from u64 to u32 with saturating_add,
reducing hash table memory per entry. Benchmark output now uses
timestamped subdirectories to preserve results across runs.

+++ src/cli.rs
@@ -197,7 +197,7 @@ pub(crate) struct Args {
 
     /// Minimum kmer count for sPCR (filters low-count kmers before PCR)
     #[arg(long, default_value_t = 2, help_heading = "General")]
-    pub(crate) min_kmer_count: u64,
+    pub(crate) min_kmer_count: u32,
 
     /// Validate FASTQ format every N records (0 = first record only)
     #[arg(long, default_value_t = 0, help_heading = "General")]
